+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+- [prettier-plugin-autocorrect](./packages/autocorrect/CHANGELOG.md)
 - [prettier-plugin-pkg](./packages/pkg/CHANGELOG.md)
 - [prettier-plugin-sh](./packages/sh/CHANGELOG.md)
 - [prettier-plugin-sql](./packages/sql/CHANGELOG.md)